@@ -18,30 +18,32 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  animated_splash_screen: ^1.0.1+2
-  buttons_tabbar: ^1.0.2+1
-  cached_network_image: ^2.4.1
-  cupertino_icons: ^1.0.0
-  dots_indicator: ^1.2.0
-  easy_localization: ^2.3.3
+  animated_splash_screen: ^1.1.0
+  buttons_tabbar: ^1.3.0
+  cached_network_image: ^3.0.0
+  cupertino_icons: ^1.0.3
+  dots_indicator: ^2.0.0
+  easy_localization: ^3.0.0
   flutter_icons: ^1.1.0
-  flutter_staggered_grid_view: ^0.3.2
-  flutter_svg: ^0.19.1
+  flutter_staggered_grid_view: ^0.4.0
+  flutter_svg: ^0.22.0
   flutter_swiper: ^1.1.6
-  fluttertoast: ^7.1.6
-  get: ^3.23.1
-  google_fonts: ^1.1.1
-  introduction_screen: ^1.0.9
-  provider: ^4.3.2+3
-  readmore: ^1.0.1
-  simple_animations: ^2.4.2
-  smooth_star_rating: ^1.1.1
+  fluttertoast: ^8.0.7
+  get: ^4.1.4
+  google_fonts: ^2.1.0
+  introduction_screen: ^2.1.0
+  provider: ^5.0.0
+  readmore: ^2.0.0
+  simple_animations: ^3.1.1
+  smooth_star_rating:
+    git:
+      url: https://github.com/educup/smoothratingbar
 
 dev_dependencies:
   flutter_test:
