@@ -5,15 +5,15 @@ class CategoryCircle extends StatelessWidget {
   final Function onTap;
 
   const CategoryCircle({
-    Key key,
-    @required this.category,
-    @required this.onTap,
+    Key? key,
+    required this.category,
+    required this.onTap,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: onTap,
+      onTap: onTap as void Function()?,
       child: Container(
         width: 80.0,
         child: Column(
@@ -26,13 +26,13 @@ class CategoryCircle extends StatelessWidget {
                 border: Border.all(
                     color: Theme.of(context).accentColor.withOpacity(.2)),
                 image: DecorationImage(
-                  image: AssetImage(category.icon),
+                  image: AssetImage(category.icon!),
                 ),
               ),
             ),
             SizedBox(height: 12.0),
             Text(
-              category.name,
+              category.name!,
               textAlign: TextAlign.center,
               maxLines: 2,
               overflow: TextOverflow.ellipsis,
