@@ -4,7 +4,7 @@ enum _AniProps { opacity, translateX }
 
 class SideInAnimation extends StatelessWidget {
   final int delay;
-  final Widget child;
+  final Widget? child;
 
   SideInAnimation(this.delay, {this.child});
 
