@@ -7,9 +7,11 @@ import 'package:dots_indicator/dots_indicator.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+// ignore: import_of_legacy_library_into_null_safe
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+// ignore: import_of_legacy_library_into_null_safe
 import 'package:flutter_swiper/flutter_swiper.dart';
 import 'package:get/get.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -46,8 +48,15 @@ part 'product/favorite_page.dart';
 part 'product/product_detail_page.dart';
 part 'product/reviews_page.dart';
 part 'product/write_review_page.dart';
-part 'profile/profile_page.dart';
+part 'profile/profile_address_page.dart';
+part 'profile/profile_change_email_page.dart';
+part 'profile/profile_change_gender_page.dart';
+part 'profile/profile_change_name_page.dart';
+part 'profile/profile_change_password_page.dart';
+part 'profile/profile_change_phone_page.dart';
 part 'profile/profile_detail_page.dart';
+part 'profile/profile_page.dart';
+part 'profile/profile_payment_page.dart';
 part 'search/filter_page.dart';
 part 'search/search.dart';
 part 'search/search_category_page.dart';
@@ -56,10 +65,3 @@ part 'search/sort_page.dart';
 part 'signIn/sign_in_page.dart';
 part 'signUp/sign_up_page.dart';
 part 'splash/splash_page.dart';
-part 'profile/profile_change_name_page.dart';
-part 'profile/profile_change_gender_page.dart';
-part 'profile/profile_change_email_page.dart';
-part 'profile/profile_change_phone_page.dart';
-part 'profile/profile_change_password_page.dart';
-part 'profile/profile_address_page.dart';
-part 'profile/profile_payment_page.dart';
